19/03/2024

+++ app/src/main/java/com/example/homeapp/data/UserData.kt
@@ -1,11 +1,17 @@
 package com.example.homeapp.data
 
+import com.google.firebase.database.Exclude
+import com.google.firebase.database.IgnoreExtraProperties
+
 class UserData(
     var userId: String? = null,
     var userName: String? = null,
     var userDate: String? = null,
     var userAddress: String? = null,
     var numberPhone: String? = null,
-    var rate: Int? = null
+    var rate: Float? = null
 ) {
+
+
+
 }
